requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
distro==1.9.0
Django==5.2.1
djangorestframework==3.16.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
openai==1.84.0
packaging==24.2
pinecone==7.0.2
pinecone-client==6.0.0
pinecone-plugin-assistant==1.6.1
pinecone-plugin-interface==0.0.7
pydantic==2.11.5
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
python-dotenv~=1.1.0
azure-core~=1.34.0
mistralai~=1.8.2